refactor MD request rate sensor in Producer/Consumer Metadata class
