+++ clients/src/main/java/org/apache/kafka/clients/consumer/KafkaConsumer.java
@@ -735,7 +735,7 @@ public class KafkaConsumer<K, V> implements Consumer<K, V> {
                     config.getLong(ConsumerConfig.METADATA_MAX_AGE_CONFIG),
                     !config.getBoolean(ConsumerConfig.EXCLUDE_INTERNAL_TOPICS_CONFIG),
                     config.getBoolean(ConsumerConfig.ALLOW_AUTO_CREATE_TOPICS_CONFIG),
-                    subscriptions, logContext, clusterResourceListeners);
+                    subscriptions, logContext, clusterResourceListeners, metrics);
             List<InetSocketAddress> addresses = ClientUtils.parseAndValidateAddresses(
                     config.getList(ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG), config.getString(ConsumerConfig.CLIENT_DNS_LOOKUP_CONFIG));
             this.metadata.bootstrap(addresses);
@@ -1014,7 +1014,7 @@ public class KafkaConsumer<K, V> implements Consumer<K, V> {
                 log.info("Subscribed to topic(s): {}", Utils.join(topics, ", "));
                 if (this.subscriptions.subscribe(new HashSet<>(topics), listener)) {
                     metadata.requestUpdateForNewTopics();
-                    kafkaConsumerMetrics.recordMetadataRequest();
+                    metadata.recordMetadataRequest();
                 }
             }
         } finally {
@@ -1080,7 +1080,7 @@ public class KafkaConsumer<K, V> implements Consumer<K, V> {
             this.subscriptions.subscribe(pattern, listener);
             this.coordinator.updatePatternSubscription(metadata.fetch());
             this.metadata.requestUpdateForNewTopics();
-            this.kafkaConsumerMetrics.recordMetadataRequest();
+            this.metadata.recordMetadataRequest();
         } finally {
             release();
         }
@@ -1173,7 +1173,7 @@ public class KafkaConsumer<K, V> implements Consumer<K, V> {
                 if (this.subscriptions.assignFromUser(new HashSet<>(partitions))) {
                     if (!skipMetadataCacheUpdate) {
                         metadata.requestUpdateForNewTopics();
-                        kafkaConsumerMetrics.recordMetadataRequest();
+                        metadata.recordMetadataRequest();
                     }
                 }
             }
