+++ clients/src/main/java/org/apache/kafka/clients/consumer/internals/ConsumerCoordinator.java
@@ -200,7 +200,7 @@ public final class ConsumerCoordinator extends AbstractCoordinator {
         }
 
         this.metadata.requestUpdate();
-        this.sensors.metadataRequestRateSensor.record();
+        this.metadata.recordMetadataRequest();
     }
 
     @Override
@@ -234,7 +234,7 @@ public final class ConsumerCoordinator extends AbstractCoordinator {
                 .collect(Collectors.toSet());
         if (subscriptions.subscribeFromPattern(topicsToSubscribe)) {
             metadata.requestUpdateForNewTopics();
-            sensors.metadataRequestRateSensor.record();
+            metadata.recordMetadataRequest();
         }
 
     }
@@ -268,7 +268,7 @@ public final class ConsumerCoordinator extends AbstractCoordinator {
 
                 if (this.subscriptions.subscribeFromPattern(newSubscription)) {
                     metadata.requestUpdateForNewTopics();
-                    sensors.metadataRequestRateSensor.record();
+                    metadata.recordMetadataRequest();
                 }
                 this.joinedSubscription = newJoinedSubscription;
             }
@@ -477,7 +477,7 @@ public final class ConsumerCoordinator extends AbstractCoordinator {
                     // passed.
                     if (this.metadata.timeToAllowUpdate(timer.currentTimeMs()) == 0) {
                         this.metadata.requestUpdate();
-                        this.sensors.metadataRequestRateSensor.record();
+                        this.metadata.recordMetadataRequest();
                     }
 
                     if (!client.ensureFreshMetadata(timer)) {
@@ -523,9 +523,11 @@ public final class ConsumerCoordinator extends AbstractCoordinator {
     private void updateGroupSubscription(Set<String> topics) {
         // the leader will begin watching for changes to any of the topics the group is interested in,
         // which ensures that all metadata changes will eventually be seen
-        if (this.subscriptions.groupSubscribe(topics))
+        if (this.subscriptions.groupSubscribe(topics)) {
             metadata.requestUpdateForNewTopics();
-        sensors.metadataRequestRateSensor.record();
+            metadata.recordMetadataRequest();
+        }
+
 
         // update metadata (if needed) and keep track of the metadata used for assignment so that
         // we can check after rebalance completion whether anything has changed
@@ -1281,19 +1283,9 @@ public final class ConsumerCoordinator extends AbstractCoordinator {
         private final Sensor assignCallbackSensor;
         private final Sensor loseCallbackSensor;
         private final Sensor pollInterval;
-        private final Sensor metadataRequestRateSensor;
 
         private ConsumerCoordinatorMetrics(Metrics metrics, String metricGrpPrefix) {
             this.metricGrpName = metricGrpPrefix + "-coordinator-metrics";
-
-            this.metadataRequestRateSensor = metrics.sensor("consumer-coordinator-metadata-request-rate");
-            this.metadataRequestRateSensor.add(new Meter(metrics.metricName("consumer-coordinator-metadata-request-rate",
-                this.metricGrpName,
-                "The average per-second number of metadata request sent by the consumer-coordinator"),
-                metrics.metricName("consumer-coordinator-metadata-request-sent-total",
-                    this.metricGrpName,
-                    "The total number of metadata requests sent by the consumer-coordinator")
-                ));
             this.commitSensor = metrics.sensor("commit-latency");
             this.commitSensor.add(metrics.metricName("commit-latency-avg",
                 this.metricGrpName,
