+++ clients/src/test/java/org/apache/kafka/clients/consumer/internals/FetcherTest.java
@@ -4180,10 +4180,10 @@ public class FetcherTest {
                                    LogContext logContext) {
         time = new MockTime(1);
         subscriptions = subscriptionState;
-        metadata = new ConsumerMetadata(0, metadataExpireMs, false, false,
-                subscriptions, logContext, new ClusterResourceListeners());
         client = new MockClient(time, metadata);
         metrics = new Metrics(metricConfig, time);
+        metadata = new ConsumerMetadata(0, metadataExpireMs, false, false,
+            subscriptions, logContext, new ClusterResourceListeners(), metrics);
         consumerClient = new ConsumerNetworkClient(logContext, client, metadata, time,
                 100, 1000, Integer.MAX_VALUE);
         metricsRegistry = new FetcherMetricsRegistry(metricConfig.tags().keySet(), "consumer" + groupId);
